+++ src/components/PeopleTable.tsx
@@ -1,0 +1,59 @@
+import { useParams } from "react-router-dom";
+import { Person } from "../types";
+import { PersonLink } from "./PersonLink";
+
+interface PeopleTableProps {
+  people: Person[];
+}
+
+export const PeopleTable = ({ people }: PeopleTableProps) => {
+  const { slug } = useParams();
+
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {people.map((person: Person) => {
+          const mother = people.find(p => p.name === person.motherName);
+          const father = people.find(p => p.name === person.fatherName);
+
+          return (
+            <tr
+              data-cy="person"
+              key={person.slug}
+              className={person.slug === slug ? 'has-background-warning' : ''}
+            >
+              <td >
+                <PersonLink person={person} />
+              </td>
+
+              <td>{person.sex}</td>
+              <td>{person.born}</td>
+              <td>{person.died}</td>
+
+              <td>
+                <PersonLink person={mother} name={person.motherName} />
+              </td>
+              <td>
+                <PersonLink person={father} name={person.fatherName} />
+              </td>
+            </tr>
+          );
+        })}
+      </tbody>
+    </table>
+  );
+};
